Rearrange all JS files, importing and exporting when nessesary. Delete unessesary files
