+++ src/battle.js
@@ -1,3 +1,5 @@
+import {PlayerCharacter, Enemy} from "./units";
+
 class BattleScene extends Phaser.Scene {
   constructor() {
     super("BattleScene");
@@ -104,14 +106,14 @@ class BattleScene extends Phaser.Scene {
   }
 
   checkEndBattle() {
-    var victory = true;
+    let victory = true;
     // if all enemies are dead we have victory
-    for (var i = 0; i < this.enemies.length; i++) {
+    for (let i = 0; i < this.enemies.length; i++) {
       if (this.enemies[i].living) victory = false;
     }
-    var gameOver = true;
+    let gameOver = true;
     // if all heroes are dead we have game over
-    for (var i = 0; i < this.heroes.length; i++) {
+    for (let i = 0; i < this.heroes.length; i++) {
       if (this.heroes[i].living) gameOver = false;
     }
     return victory || gameOver;
@@ -121,7 +123,7 @@ class BattleScene extends Phaser.Scene {
     // clear state, remove sprites
     this.heroes.length = 0;
     this.enemies.length = 0;
-    for (var i = 0; i < this.units.length; i++) {
+    for (let i = 0; i < this.units.length; i++) {
       // link item
       this.units[i].destroy();
     }
@@ -153,7 +155,7 @@ class BattleScene extends Phaser.Scene {
     } else {
       // else if its enemy unit
       // pick random living hero to be attacked
-      var r;
+      let r;
       do {
         r = Math.floor(Math.random() * this.heroes.length);
       } while (!this.heroes[r].living);
@@ -169,339 +171,9 @@ class BattleScene extends Phaser.Scene {
   }
 }
 
-class UIScene extends Phaser.Scene {
-  constructor() {
-    super("UIScene");
-  }
-  preload() {}
-
-  remapHeroes() {
-    let heroes = this.battleScene.heroes;
-    this.heroesMenu.remap(heroes);
-  }
-
-  remapEnemies() {
-    var enemies = this.battleScene.enemies;
-    this.enemiesMenu.remap(enemies);
-  }
-
-  onKeyInput(event) {
-    if (this.currentMenu && this.currentMenu.selected) {
-      if (event.code === "ArrowUp") {
-        this.currentMenu.moveSelectionUp();
-      } else if (event.code === "ArrowDown") {
-        this.currentMenu.moveSelectionDown();
-      } else if (event.code === "ArrowRight" || event.code === "Shift") {
-      } else if (event.code === "Space" || event.code === "ArrowLeft") {
-        this.currentMenu.confirm();
-      }
-    }
-  }
-
-  onPlayerSelect(id) {
-    this.heroesMenu.select(id);
-    this.actionsMenu.select(0);
-    this.currentMenu = this.actionsMenu;
-  }
-
-  onSelectAction() {
-    this.currentMenu = this.enemiesMenu;
-    this.enemiesMenu.select(0);
-  }
-
-  onEnemy(index) {
-    this.heroesMenu.deselect();
-    this.actionsMenu.deselect();
-    this.enemiesMenu.deselect();
-    this.currentMenu = null;
-    this.battleScene.receivePlayerSelection("attack", index);
-  }
-
-  createMenu() {
-    // map hero menu items to heroes
-    this.remapHeroes();
-    // map enemies menu items to enemies
-    this.remapEnemies();
-    // first move
-    this.battleScene.nextTurn();
-  }
-
-  create() {
-    this.graphics = this.add.graphics();
-    this.graphics.lineStyle(1, 0xffffff);
-    this.graphics.fillStyle(0x031f4c, 1);
-    this.graphics.strokeRect(2, 400, 220, 200);
-    this.graphics.fillRect(2, 400, 220, 200);
-    this.graphics.strokeRect(225, 400, 220, 200);
-    this.graphics.fillRect(225, 400, 220, 200);
-    this.graphics.strokeRect(448, 400, 350, 200);
-    this.graphics.fillRect(448, 400, 350, 200);
-
-    // basic container to hold all menus
-    this.menus = this.add.container();
-
-    this.heroesMenu = new HeroesMenu(460, 425, this);
-    this.actionsMenu = new ActionsMenu(250, 425, this);
-    this.enemiesMenu = new EnemiesMenu(8, 425, this);
-
-    // the currently selected menu
-
-    this.currentMenu = this.actionsMenu;
-
-    // add menus to the container
-    this.menus.add(this.heroesMenu);
-    this.menus.add(this.actionsMenu);
-    this.menus.add(this.enemiesMenu);
-
-    this.battleScene = this.scene.get("BattleScene");
-
-    // listen for keyboard events
-    this.input.keyboard.on("keydown", this.onKeyInput, this);
-
-    // when its player cunit turn to move
-    this.battleScene.events.on("PlayerSelect", this.onPlayerSelect, this);
-
-    // when the action on the menu is selected
-    // for now we have only one action so we dont send and action id
-    this.events.on("SelectAction", this.onSelectAction, this);
-
-    // an enemy is selected
-    this.events.on("Enemy", this.onEnemy, this);
-
-    // when the scene receives wake event
-    this.sys.events.on("wake", this.createMenu, this);
-
-    // the message describing the current action
-    this.message = new Message(this, this.battleScene.events);
-    this.add.existing(this.message);
-
-    this.createMenu();
-  }
-}
-
-// ************************************************************** UNIT CREATORS
-
-class Unit extends Phaser.GameObjects.Sprite {
-  constructor(scene, x, y, texture, frame, type, hp, damage) {
-    super(scene, x, y, texture, frame);
-    this.type = type;
-    this.maxHp = this.hp = hp;
-    this.damage = damage; // default damage
-    this.living = true;
-    this.menuItem = null;
-  }
-
-  setMenuItem(item) {
-    this.menuItem = item;
-  }
-
-  attack(target) {
-    if (target.living) {
-      target.takeDamage(this.damage);
-      this.scene.events.emit(
-        "Message",
-        this.type +
-          " attacks " +
-          target.type +
-          " for " +
-          this.damage +
-          " damage"
-      );
-    }
-  }
-  takeDamage(damage) {
-    this.hp -= damage;
-    if (this.hp <= 0) {
-      this.hp = 0;
-      this.menuItem.unitKilled();
-      this.living = false;
-      this.visible = false;
-      this.menuItem = null;
-    }
-  }
-}
-
-class Enemy extends Unit {
-  constructor(scene, x, y, texture, frame, type, hp, damage) {
-    super(scene, x, y, texture, frame, type, hp, damage);
-    this.setScale(2);
-  }
-}
-
-class PlayerCharacter extends Unit {
-  constructor(scene, x, y, texture, frame, type, hp, damage) {
-    super(scene, x, y, texture, frame, type, hp, damage);
-    // flip the image so I don't have to edit it manually
-    this.flipX = true;
-
-    this.setScale(1.5);
-  }
-}
-
-// ***************************************************************MENU ITEMS AND BODY
-
-class MenuItem extends Phaser.GameObjects.Text {
-  constructor(x, y, text, scene) {
-    super(scene, x, y, text, { color: "#ffffff", align: "left", fontSize: 15 });
-  }
-  select() {
-    this.setColor("#f8ff38");
-  }
-  deselect() {
-    this.setColor("#ffffff");
-  }
-  unitKilled() {
-    this.active = false;
-    this.visible = false;
-  }
-}
-
-class Menu extends Phaser.GameObjects.Container {
-  constructor(x, y, scene, heroes) {
-    super(scene, x, y);
-    this.menuItems = [];
-    this.menuItemIndex = 0;
-    this.heroes = heroes;
-    this.x = x;
-    this.y = y;
-    this.selected = false;
-  }
-
-  addMenuItem(unit) {
-    var menuItem = new MenuItem(
-      0,
-      this.menuItems.length * 20,
-      unit,
-      this.scene
-    );
-    this.menuItems.push(menuItem);
-    this.add(menuItem);
-    return menuItem;
-  }
-
-  moveSelectionUp() {
-    this.menuItems[this.menuItemIndex].deselect();
-    do {
-      this.menuItemIndex--;
-      if (this.menuItemIndex < 0)
-        this.menuItemIndex = this.menuItems.length - 1;
-    } while (!this.menuItems[this.menuItemIndex].active);
-    this.menuItems[this.menuItemIndex].select();
-  }
-
-  moveSelectionDown() {
-    this.menuItems[this.menuItemIndex].deselect();
-    do {
-      this.menuItemIndex++;
-      if (this.menuItemIndex >= this.menuItems.length) this.menuItemIndex = 0;
-    } while (!this.menuItems[this.menuItemIndex].active);
-    this.menuItems[this.menuItemIndex].select();
-  }
-
-  select(index) {
-    if (!index) index = 0;
-    this.menuItems[this.menuItemIndex].deselect();
-    this.menuItemIndex = index;
-    while (!this.menuItems[this.menuItemIndex].active) {
-      this.menuItemIndex++;
-      if (this.menuItemIndex >= this.menuItems.length) this.menuItemIndex = 0;
-      if (this.menuItemIndex == index) return;
-    }
-    this.menuItems[this.menuItemIndex].select();
-    this.selected = true;
-  }
-
-  deselect() {
-    this.menuItems[this.menuItemIndex].deselect();
-    this.menuItemIndex = 0;
-    this.selected = false;
-  }
-
-  confirm() {}
-
-  clear() {
-    for (var i = 0; i < this.menuItems.length; i++) {
-      this.menuItems[i].destroy();
-    }
-    this.menuItems.length = 0;
-    this.menuItemIndex = 0;
-  }
-
-  remap(units) {
-    this.clear();
-    for (var i = 0; i < units.length; i++) {
-      var unit = units[i];
-      unit.setMenuItem(this.addMenuItem(unit.type));
-    }
-    this.menuItemIndex = 0;
-  }
-}
-
-class HeroesMenu extends Menu {
-  constructor(x, y, scene) {
-    super(x, y, scene);
-  }
-}
-
-class ActionsMenu extends Menu {
-  constructor(x, y, scene) {
-    super(x, y, scene);
-    this.addMenuItem("Attack");
-  }
-
-  confirm() {
-    this.scene.events.emit("SelectAction");
-  }
-}
-
-class EnemiesMenu extends Menu {
-  constructor(x, y, scene) {
-    super(x, y, scene);
-  }
-
-  confirm() {
-    this.scene.events.emit("Enemy", this.menuItemIndex);
-  }
-}
 
 // ************************************************* MESSAGES
 
-class Message extends Phaser.GameObjects.Container {
-  constructor(scene, events) {
-    super(scene, 260, 40);
-    const graphics = this.scene.add.graphics();
-    this.add(graphics);
-    graphics.lineStyle(1, 0xffffff, 0.8);
-    graphics.fillStyle(0x031f4c, 0.3);
-    graphics.strokeRect(-145, -25, 280, 60);
-    graphics.fillRect(-145, -25, 280, 60);
-    this.text = new Phaser.GameObjects.Text(scene, 0, 0, "", {
-      color: "#ffffff",
-      align: "center",
-      fontSize: 17,
-      wordWrap: { width: 260, useAdvancedWrap: true },
-    });
-    this.add(this.text);
-    this.text.setOrigin(0.5);
-    events.on("Message", this.showMessage, this);
-    this.visible = false;
-  }
 
-  showMessage(text) {
-    this.text.setText(text);
-    this.visible = true;
-    if (this.hideEvent) this.hideEvent.remove(false);
-    this.hideEvent = this.scene.time.addEvent({
-      delay: 2000,
-      callback: this.hideMessage,
-      callbackScope: this,
-    });
-  }
 
-  hideMessage() {
-    this.hideEvent = null;
-    this.visible = false;
-  }
-}
-
-export { BattleScene, UIScene };
+export { BattleScene};
