--- src/introduction.js
@@ -1,6 +1,0 @@
-export class Introduction extends Phaser.Scene {
-  constructor() {
-    super("Introduction");
-  }
-
-}
